[FEAT] 숫자야구 답안 제출기록 보이게 수정, 랭킹 기능구현

+++ baseball-game-web/src/main/java/jk/baseballgameweb/rank/Rank.java
@@ -20,12 +20,12 @@ public class Rank {
 
     @Enumerated(EnumType.STRING)
     private GameType game;
-    private Long point;
+    private int point;
 
     public Rank() {
     }
 
-    public Rank(Member member, GameType game, Long point) {
+    public Rank(Member member, GameType game, int point) {
         this.member = member;
         this.game = game;
         this.point = point;
